refactor(api): move organization-administration controller
to src/organizational-entities
